+++ Seven/Structures/List.cs
@@ -57,7 +57,7 @@ namespace Seven.Structures
     
     /// <summary>Gets the current memory imprint of this structure in bytes.</summary>
     /// <remarks>Returns long.MaxValue on overflow.</remarks>
-    public long SizeOf { get { return _count; } }
+    public int SizeOf { get { return _count; } }
 
     /// <summary>Returns the number of items in the list.</summary>
     /// <remarks>Runtime: O(1).</remarks>
@@ -578,7 +578,7 @@ namespace Seven.Structures
 
     /// <summary>Gets the current memory imprint of this structure in bytes.</summary>
     /// <remarks>Returns long.MaxValue on overflow.</remarks>
-    public long SizeOf { get { return this._list.Length; } }
+    public int SizeOf { get { return this._list.Length; } }
 
     /// <summary>Pulls out all the values in the structure that are equivalent to the key.</summary>
     /// <typeparam name="Key">The type of the key to check for.</typeparam>
@@ -801,7 +801,7 @@ namespace Seven.Structures
 
     /// <summary>Gets the current memory imprint of this structure in bytes.</summary>
     /// <remarks>Returns long.MaxValue on overflow.</remarks>
-    public long SizeOf { get { return this._count; } }
+    public int SizeOf { get { return this._count; } }
 
     /// <summary>Pulls out all the values in the structure that are equivalent to the key.</summary>
     /// <typeparam name="Key">The type of the key to check for.</typeparam>
