+++ Seven/Structures/Stack.cs
@@ -139,7 +139,7 @@ namespace Seven.Structures
 
     /// <summary>Gets the current memory imprint of this structure in bytes.</summary>
     /// <remarks>Returns long.MaxValue on overflow.</remarks>
-    public long SizeOf { get { return this._count; } }
+    public int SizeOf { get { return this._count; } }
 
     /// <summary>Pulls out all the values in the structure that are equivalent to the key.</summary>
     /// <typeparam name="Key">The type of the key to check for.</typeparam>
@@ -536,7 +536,7 @@ namespace Seven.Structures
 
     /// <summary>Gets the current memory imprint of this structure in bytes.</summary>
     /// <remarks>Returns long.MaxValue on overflow.</remarks>
-    public long SizeOf { get { return this._count; } }
+    public int SizeOf { get { return this._count; } }
 
     /// <summary>Pulls out all the values in the structure that are equivalent to the key.</summary>
     /// <typeparam name="Key">The type of the key to check for.</typeparam>
